requirements.txt
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
docker==7.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
librt==0.7.8
mypy==1.19.1
mypy_extensions==1.1.0
notion-client==2.2.1
pathspec==1.0.3
pycparser==3.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
types-docker==7.1.0.20260109
types-paramiko==4.0.0.20250822
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.15.0
urllib3==2.6.3
